+++ ChangeLog.md
@@ -1,5 +1,11 @@
 ## change log ##
 
+* 2015-12-09: orrportal 2.7.x - \#361 Preparations for "build and test portal version against orr-ont endpoint"
+  - Refactoring to extract some general stuff from OrrClientImpl into new OrrClientImplBase 
+  - New OrrClientImpl2 for the implementation against orr-ont (still incomplete)
+  - OrrClientFactory: ad hoc choice of implementation during initial impl/testing: if the ont.service.url
+    config property ends with "/orr-ont" then the new implementation is used.
+  
 * 2015-12-08:
   Re \#364 "dispatch portal functionality through ont-based URL directly"
   Although the general desired behavior is already implemented, here are some observations that make the